Move creating slack users into EM.run block.
We need to be sure the event loop is configured and starting before
trying to defer a block. If deferred outside the #run block, we can get
`a eventmachine not initialized: evma_signal_loopbreak` error if we try
to stop the event loop before it's started.

We might also consider stubbing and yielding the `EventLoop.defer`
method in a general `before` block for this entire set of specs, to make
sure we're not accidentally deferring work we don't intend to.