Properly set the room name in received messages
Before, the channel ID was being set as both the `id` and `name` within
the message's `room_object`. This will either lookup the correct room
object, which would have a name, or it will assign the channel ID as a
fallback (such as when it's a private conversation without a name).