Move where block is yielded at connection
The block was yielded at the connection open because we wanted messages
to only be processed when the websocket was already open. All messages
are sent via the API anyway, we just go through the RTMConnection object
to get the user->channel mapping, which is already populated by
tm_start data and doesn't depend on the websockets connection being up.

This will also allow private messages to be sent without issues
even when the RTM connection is not up anymore.